requirements.txt
flask
flask-bootstrap
arrow
boto3
botocore
requests
Wand
pyheif
Pillow
pydantic